refactor: fix component media URLs
